fix: Fix display of download progress
The floating point conversion was done too late.

+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/main/transferdetails/TransferDownloadComposeUi.kt
@@ -302,7 +302,7 @@ class TransferDownloadComposeUi(
 }
 
 private fun InProgress.getProgress(): Float {
-    return (if (totalSizeInBytes == 0L) 0 else downloadedBytes / totalSizeInBytes).toFloat()
+    return (if (totalSizeInBytes == 0L) 0 else downloadedBytes.toDouble() / totalSizeInBytes).toFloat()
 }
 
 @OptIn(ExperimentalPermissionsApi::class)
